requirements.txt
pillow
pillow-avif-plugin
retina-face
opencv-python
dateparser